Refactor. Simplification of the structure.
