+++ app.js
@@ -2,17 +2,17 @@ const express = require('express')
 const logger = require('morgan')
 const cors = require('cors')
 
-const indexRouter = require('./routes/index')
+const contactsRouter = require('./routes/api/contacts')
 
 const app = express()
 
-const formatsLogger = app.get('env') === 'development' ? 'dev' : 'combined'
+const formatsLogger = app.get('env') === 'development' ? 'dev' : 'short'
 
 app.use(logger(formatsLogger))
 app.use(cors())
 app.use(express.json())
 
-app.use('/api/contacts', indexRouter)
+app.use('/api/contacts', contactsRouter)
 
 app.use((req, res) => {
   res.status(404).json({ message: 'Not found' })
