+++ bin/server.js
@@ -1,53 +1,7 @@
-#!/usr/bin/env node
-
 const app = require('../app')
-const debug = require('debug')('template:server')
-const http = require('http')
 
-const normalizePort = (val) => {
-  const port = parseInt(val, 10)
+const PORT = process.env.PORT || 3000
 
-  if (isNaN(port)) {
-    return val
-  }
-
-  if (port >= 0) {
-    return port
-  }
-
-  return false
-}
-
-const port = normalizePort(process.env.PORT || '3000')
-app.set('port', port)
-
-const server = http.createServer(app)
-
-server.listen(port)
-server.on('error', onError)
-server.on('listening', onListening)
-
-function onError(error) {
-  if (error.syscall !== 'listen') {
-    throw error
-  }
-
-  const bind = typeof port === 'string' ? 'Pipe ' + port : 'Port ' + port
-
-  switch (error.code) {
-    case 'EACCES':
-      console.error(bind + ' requires elevated privileges')
-      process.exit(1)
-    case 'EADDRINUSE':
-      console.error(bind + ' is already in use')
-      process.exit(1)
-    default:
-      throw error
-  }
-}
-
-function onListening() {
-  const addr = server.address()
-  const bind = typeof addr === 'string' ? 'pipe ' + addr : 'port ' + addr.port
-  debug('Listening on ' + bind)
-}
+app.listen(PORT, () => {
+  console.log(`Server running. Use our API on port: ${PORT}`)
+})
