+++ package.json
@@ -4,15 +4,13 @@
   "private": true,
   "scripts": {
     "start": "cross-env NODE_ENV=production node ./bin/server.js",
-    "start:dev": "nodemon ./bin/server.js",
-    "start:debug": "cross-env DEBUG=* nodemon ./bin/server.js",
+    "start:dev": "cross-env NODE_ENV=development nodemon ./bin/server.js",
     "lint": "eslint **/*.js",
     "lint:fix": "eslint --fix **/*.js"
   },
   "dependencies": {
     "cors": "2.8.5",
     "cross-env": "7.0.3",
-    "debug": "4.3.1",
     "express": "4.17.1",
     "morgan": "1.10.0"
   },
